+++ packages/data-model/src/entity/AcDbSpline.ts
@@ -20,39 +20,52 @@ export class AcDbSpline extends AcDbCurve {
    * @param knots Input an array of numbers that specifies the knot values of the spline
    * @param weights Input an array of doubles that specifies the weights at each control point.
    * Default weight of 1 if weights are not provided
+   * @param closed Whether the spline should be closed. Default is false
    */
   constructor(
     controlPoints: AcGePointLike[],
     knots: number[],
-    weights?: number[]
+    weights?: number[],
+    closed?: boolean
   )
   /**
    * Construct an instance of the spline entity.
    * @param fitPoints Input an array of points (in WCS coordinates) through which to fit the curve
    * @param knotParam Input knot parameterization which define the knot values
+   * @param closed Whether the spline should be closed. Default is false
    */
   constructor(
     fitPoints: AcGePointLike[],
-    knotParam: AcGeKnotParameterizationType
+    knotParam: AcGeKnotParameterizationType,
+    closed?: boolean
   )
-  constructor(a?: unknown, b?: unknown, c?: unknown) {
+  constructor(a?: unknown, b?: unknown, c?: unknown, d?: unknown) {
     super()
     const argsLength =
-      +(a !== undefined) + +(b !== undefined) + +(c !== undefined)
+      +(a !== undefined) +
+      +(b !== undefined) +
+      +(c !== undefined) +
+      +(d !== undefined)
 
-    if (argsLength != 2 && argsLength != 3) {
+    if (argsLength < 2 || argsLength > 4) {
       throw AcCmErrors.ILLEGAL_PARAMETERS
     }
-    if (argsLength == 2 && !Array.isArray(b)) {
+
+    // Determine if this is the fitPoints constructor (2 or 3 args, second arg is not an array)
+    const isFitPointsConstructor = argsLength <= 3 && !Array.isArray(b)
+
+    if (isFitPointsConstructor) {
       this._geo = new AcGeSpline3d(
         a as AcGePointLike[],
-        b as AcGeKnotParameterizationType
+        b as AcGeKnotParameterizationType,
+        c as boolean
       )
     } else {
       this._geo = new AcGeSpline3d(
         a as AcGePointLike[],
         b as number[],
-        c as number[] | undefined
+        c as number[] | undefined,
+        d as boolean
       )
     }
   }