+++ packages/libredwg-converter/src/AcDbEntitiyConverter.ts
@@ -184,17 +184,14 @@ export class AcDbEntityConverter {
 
   private convertSpline(spline: DwgSplineEntity) {
     if (spline.numberOfControlPoints > 0 && spline.numberOfKnots > 0) {
-      const dbEntity = new AcDbSpline(
+      return new AcDbSpline(
         spline.controlPoints,
         spline.knots,
-        spline.weights
+        spline.weights,
+        !!(spline.flag & 0x01)
       )
-      dbEntity.closed = !!(spline.flag & 0x01)
-      return dbEntity
     } else if (spline.numberOfFitPoints > 0) {
-      const dbEntity = new AcDbSpline(spline.fitPoints, 'Uniform')
-      dbEntity.closed = !!(spline.flag & 0x01)
-      return dbEntity
+      return new AcDbSpline(spline.fitPoints, 'Uniform', !!(spline.flag & 0x01))
     }
     return null
   }